+++ crawlserv/src/Module/Thread.hpp
@@ -52,7 +52,7 @@
 #include <atomic>				// std::atomic
 #include <chrono>				// std::chrono
 #include <cmath>				// std::lround
-#include <cstddef>				// std::size_t
+#include <cstdint>				// std::uint64_t
 #include <condition_variable>	// std::condition_variable
 #include <exception>			// std::exception
 #include <iostream>				// std::cout, std::flush
@@ -91,10 +91,10 @@ namespace crawlservpp::Module {
 		virtual ~Thread();
 
 		// getters
-		std::size_t getId() const;
-		std::size_t getWebsite() const;
-		std::size_t getUrlList() const;
-		std::size_t getConfig() const;
+		std::uint64_t getId() const;
+		std::uint64_t getWebsite() const;
+		std::uint64_t getUrlList() const;
+		std::uint64_t getConfig() const;
 		bool isShutdown() const;
 		bool isRunning() const;
 		bool isFinished() const;
@@ -109,7 +109,7 @@ namespace crawlservpp::Module {
 		void end();
 
 		// time travel
-		void warpTo(std::size_t target);
+		void warpTo(std::uint64_t target);
 
 		// class for Thread exceptions
 		MAIN_EXCEPTION_CLASS();
@@ -140,8 +140,8 @@ namespace crawlservpp::Module {
 		void allowPausing();
 		void disallowPausing();
 
-		std::size_t getLast() const;
-		void setLast(std::size_t last);
+		std::uint64_t getLast() const;
+		void setLast(std::uint64_t last);
 		void incrementLast();
 		std::string getStatusMessage() const;
 		long getWarpedOverAndReset();
@@ -153,38 +153,38 @@ namespace crawlservpp::Module {
 		virtual void onClear() = 0;
 
 	private:
-		Main::Database& databaseClass;			// access to the database for the class
+		Main::Database& databaseClass;				// access to the database for the class
 
-		std::atomic<bool> pausable;				// thread is pausable
- 		std::atomic<bool> running;				// thread is running (or paused)
-		std::atomic<bool> paused;				// thread is paused
-		std::atomic<bool> interrupted;			// thread has been interrupted by shutdown
-		std::atomic<bool> terminated;			// thread has been terminated due to an exception
-		std::atomic<bool> shutdown;				// shutdown in progress
-		std::atomic<bool> finished;				// shutdown is finished
+		std::atomic<bool> pausable;					// thread is pausable
+ 		std::atomic<bool> running;					// thread is running (or paused)
+		std::atomic<bool> paused;					// thread is paused
+		std::atomic<bool> interrupted;				// thread has been interrupted by shutdown
+		std::atomic<bool> terminated;				// thread has been terminated due to an exception
+		std::atomic<bool> shutdown;					// shutdown in progress
+		std::atomic<bool> finished;					// shutdown is finished
 
-		std::size_t id;							// the ID of the thread in the database
-		std::string module;						// the module of the thread (used for logging)
-		ThreadOptions options;					// options for the thread
+		std::uint64_t id;							// the ID of the thread in the database
+		std::string module;							// the module of the thread (used for logging)
+		ThreadOptions options;						// options for the thread
 
-		std::size_t last;						// last ID for the thread
-		std::atomic<std::size_t> overwriteLast;	// ID to overwrite last ID with ("time travel")
-		long warpedOver;						// no. of IDs that have been skipped (might be negative, ONLY for threads!)
+		std::uint64_t last;							// last ID for the thread
+		std::atomic<std::uint64_t> overwriteLast;	// ID to overwrite last ID with ("time travel")
+		long warpedOver;							// no. of IDs that have been skipped (might be negative, ONLY for threads!)
 
-		std::condition_variable pauseCondition;	// condition variable to wait for unpause
-		mutable std::mutex pauseLock;			// lock for accessing the condition variable
+		std::condition_variable pauseCondition;		// condition variable to wait for unpause
+		mutable std::mutex pauseLock;				// lock for accessing the condition variable
 
-		std::string status; 					// status message of the thread (without pause state)
-		mutable std::mutex statusLock;			// lock for accessing the status message
+		std::string status; 						// status message of the thread (without pause state)
+		mutable std::mutex statusLock;				// lock for accessing the status message
 
-		std::thread thread;						// pointer to the thread
+		std::thread thread;							// pointer to the thread
 
 		// timing statistics (in seconds)
 		std::chrono::steady_clock::time_point startTimePoint;
 		std::chrono::steady_clock::time_point pauseTimePoint;
-		std::chrono::duration<std::size_t> runTime;
-		std::chrono::duration<std::size_t> pauseTime;
-		std::size_t getRunTime() const;
+		std::chrono::duration<std::uint64_t> runTime;
+		std::chrono::duration<std::uint64_t> pauseTime;
+		std::uint64_t getRunTime() const;
 		void updateRunTime();
 		void updatePauseTime();
